# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem:
    properties:
      code:
        type: string
      detail:
        type: string
      meta:
        additionalProperties: {}
        type: object
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.AdoptReceipt:
    properties:
      changed:
        type: boolean
      duplicate:
        type: boolean
      episode_mappings:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeMapping'
        type: array
      previous_version_id:
        type: string
      project_revision:
        type: integer
      script_revision:
        type: integer
      split_set_id:
        type: string
      version_id:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.ConfirmationPage:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.ConfirmationSummary'
        type: array
      next_revision:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.ConfirmationSummary:
    properties:
      candidate_set_id:
        type: string
      created_at:
        type: string
      episode_count:
        type: integer
      formal_set_id:
        type: string
      id:
        type: string
      revision:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeMapping:
    properties:
      episode_id:
        type: string
      inherit_status:
        type: string
      previous_episode_id:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceControlReceipt:
    properties:
      accepted:
        type: boolean
      action:
        type: string
      intent_id:
        type: string
      revision:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceWriteIntent:
    properties:
      action:
        type: string
      active_io:
        type: boolean
      base_version_id:
        type: string
      can_control:
        type: boolean
      cancellation_requested:
        type: boolean
      confirmed_object_count:
        type: integer
      created_at:
        type: string
      expected_script_revision:
        type: integer
      id:
        type: string
      needs_reconciliation:
        type: boolean
      object_count:
        type: integer
      revision:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceWritePage:
    properties:
      current_actor_id:
        type: string
      current_org_id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceWriteIntent'
        type: array
      next_after:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionText:
    properties:
      content_hash:
        type: string
      script_version_id:
        type: string
      span_end:
        type: integer
      span_start:
        type: integer
      text:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode:
    properties:
      confirmed_structure_id:
        type: string
      current_structure_id:
        type: string
      id:
        type: string
      inherit_status:
        type: string
      is_delete:
        type: boolean
      org_id:
        type: string
      previous_episode_id:
        type: string
      project_id:
        type: string
      revision:
        type: integer
      script_version_id:
        type: string
      seq_no:
        type: integer
      span_end:
        type: integer
      span_start:
        type: integer
      split_set_id:
        type: string
      title:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.ScalarSpan:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.SplitConfirmation:
    properties:
      actor_id:
        type: string
      candidate_set_id:
        type: string
      created_at:
        type: string
      episodes:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode'
        type: array
      formal_set_id:
        type: string
      id:
        type: string
      org_id:
        type: string
      preface:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.ScalarSpan'
      project_id:
        type: string
      revision:
        type: integer
      version_id:
        type: string
    type: object
...
  internal_script_adapter_http.AdoptVersionRequest:
    properties:
      ack_invalidate:
        type: boolean
      expected_revision:
        type: integer
      expected_split_revision:
        type: integer
    type: object
...
  internal_script_adapter_http.SourceControlRequest:
    properties:
      expected_revision:
        type: integer
    type: object
...
info:
  contact: {}
  description: Single workspace project query, resource canvas commands and authorized
    media previews.
  title: Lanverse API
  version: "0.1"
paths:
...
  /api/projects/{pid}/script-source-writes:
    get:
      operationId: listScriptSourceWrites
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 分页位置
        in: query
        name: after
        type: integer
      - description: 1..100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceWritePage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 未完成正文保存的安全分页
      tags:
      - script
  /api/projects/{pid}/script-source-writes/{wid}:
    get:
      operationId: getScriptSourceWrite
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 保存意图UUID
        in: path
        name: wid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceWriteIntent'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 正文保存意图状态
      tags:
      - script
  /api/projects/{pid}/script-source-writes/{wid}/cancel:
    post:
      consumes:
      - application/json
      operationId: cancelScriptSourceWrite
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 保存意图UUID
        in: path
        name: wid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 追踪UUID
        in: header
        name: X-Request-Id
        type: string
      - description: 保存意图修订
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.SourceControlRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceControlReceipt'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 明确取消未发布正文保存
      tags:
      - script
  /api/projects/{pid}/script-source-writes/{wid}/reconcile:
    post:
      consumes:
      - application/json
      operationId: reconcileScriptSourceWrite
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 保存意图UUID
        in: path
        name: wid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 追踪UUID
        in: header
        name: X-Request-Id
        type: string
      - description: 保存意图修订
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.SourceControlRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceControlReceipt'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 核验原正文保存对象与停止证据
      tags:
      - script
...
  /api/projects/{pid}/script-versions/{vid}/adopt:
    post:
      consumes:
      - application/json
      operationId: adoptScriptVersion
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 不可变剧本版本UUID
        in: path
        name: vid
        required: true
        type: string
      - description: 永久UUID请求键
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 头修订及影响确认
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.AdoptVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.AdoptReceipt'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 采纳已确认分集的正式剧本版本
      tags:
      - script
  /api/projects/{pid}/script-versions/{vid}/source-text:
    get:
      operationId: getScriptVersionSourceText
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 不可变剧本版本UUID
        in: path
        name: vid
        required: true
        type: string
      - description: 包含起点，Unicode scalar
        in: query
        name: from
        required: true
        type: integer
      - description: 不含终点，最多65536scalar
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionText'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 候选与历史版本的规范正文片段
      tags:
      - script
  /api/projects/{pid}/script-versions/{vid}/split-confirmations:
    get:
      operationId: listScriptSplitConfirmations
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 不可变剧本版本UUID
        in: path
        name: vid
        required: true
        type: string
      - description: 严格早于确认修订
        in: query
        name: before_revision
        type: integer
      - description: 1..100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.ConfirmationPage'
      summary: 分集确认的不可变历史
      tags:
      - script
  /api/projects/{pid}/script-versions/{vid}/split-confirmations/{cid}:
    get:
      operationId: getScriptSplitConfirmation
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 不可变剧本版本UUID
        in: path
        name: vid
        required: true
        type: string
      - description: 确认UUID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.SplitConfirmation'
      summary: 指定分集确认的完整边界与身份结果
      tags:
      - script
...
swagger: "2.0"